+++ frontends/ol-widgets/src/components/RichTextWidgetContent.tsx
@@ -1,9 +1,9 @@
 import React from "react"
 import type { RichTextWidgetInstance } from "../interfaces"
 
-const RichTextWdigetContent: React.FC<{ widget: RichTextWidgetInstance }> = ({
-  widget
-}) => {
+const RichTextWdigetContent: React.FC<{
+  widget: Omit<RichTextWidgetInstance, "id">
+}> = ({ widget }) => {
   return <>{widget.configuration.source}</>
 }
